+++ CHANGELOG.md
@@ -53,8 +53,7 @@ power during expiry
 ### State Machine Breaking
 
 * [#224](https://github.com/babylonlabs-io/babylon/pull/224) Make injected checkpoint a standard tx
-* [#216](https://github.com/babylonlabs-io/babylon/pull/216) move voting power distribution
-update algorithm to `x/finality`
+* [#217](https://github.com/babylonlabs-io/babylon/pull/217) Move voting power table to finality module
 * [#207](https://github.com/babylonlabs-io/babylon/pull/207) Rename total voting power
 to total bonded sat
 * [#204](https://github.com/babylonlabs-io/babylon/pull/204) Add babylon finality
