+++ client/query/btcstaking.go
@@ -121,48 +121,3 @@ func (c *QueryClient) BTCDelegation(stakingTxHashHex string) (*btcstakingtypes.Q
 
 	return resp, err
 }
-
-// ActiveFinalityProvidersAtHeight queries the BTCStaking module for all finality providers
-// with non-zero voting power at a given height
-func (c *QueryClient) ActiveFinalityProvidersAtHeight(height uint64, pagination *sdkquerytypes.PageRequest) (*btcstakingtypes.QueryActiveFinalityProvidersAtHeightResponse, error) {
-	var resp *btcstakingtypes.QueryActiveFinalityProvidersAtHeightResponse
-	err := c.QueryBTCStaking(func(ctx context.Context, queryClient btcstakingtypes.QueryClient) error {
-		var err error
-		req := &btcstakingtypes.QueryActiveFinalityProvidersAtHeightRequest{
-			Height:     height,
-			Pagination: pagination,
-		}
-		resp, err = queryClient.ActiveFinalityProvidersAtHeight(ctx, req)
-		return err
-	})
-
-	return resp, err
-}
-
-// FinalityProviderPowerAtHeight queries the BTCStaking module for the power of a finality provider at a given height
-func (c *QueryClient) FinalityProviderPowerAtHeight(fpBtcPkHex string, height uint64) (*btcstakingtypes.QueryFinalityProviderPowerAtHeightResponse, error) {
-	var resp *btcstakingtypes.QueryFinalityProviderPowerAtHeightResponse
-	err := c.QueryBTCStaking(func(ctx context.Context, queryClient btcstakingtypes.QueryClient) error {
-		var err error
-		req := &btcstakingtypes.QueryFinalityProviderPowerAtHeightRequest{
-			FpBtcPkHex: fpBtcPkHex,
-			Height:     height,
-		}
-		resp, err = queryClient.FinalityProviderPowerAtHeight(ctx, req)
-		return err
-	})
-
-	return resp, err
-}
-
-func (c *QueryClient) ActivatedHeight() (*btcstakingtypes.QueryActivatedHeightResponse, error) {
-	var resp *btcstakingtypes.QueryActivatedHeightResponse
-	err := c.QueryBTCStaking(func(ctx context.Context, queryClient btcstakingtypes.QueryClient) error {
-		var err error
-		req := &btcstakingtypes.QueryActivatedHeightRequest{}
-		resp, err = queryClient.ActivatedHeight(ctx, req)
-		return err
-	})
-
-	return resp, err
-}
