+++ proto/babylon/btcstaking/v1/genesis.proto
@@ -4,7 +4,6 @@ package babylon.btcstaking.v1;
 import "gogoproto/gogo.proto";
 import "babylon/btcstaking/v1/params.proto";
 import "babylon/btcstaking/v1/btcstaking.proto";
-import "babylon/btcstaking/v1/incentive.proto";
 import "babylon/btcstaking/v1/events.proto";
 
 option go_package = "github.com/babylonlabs-io/babylon/x/btcstaking/types";
@@ -17,36 +16,12 @@ message GenesisState {
   repeated FinalityProvider finality_providers = 2;
   // btc_delegations all the btc delegations in the state.
   repeated BTCDelegation btc_delegations = 3;
-  // voting_powers the voting power of every finality provider at every block height.
-  repeated VotingPowerFP voting_powers = 4;
   // block_height_chains the block height of babylon and bitcoin.
   repeated BlockHeightBbnToBtc block_height_chains = 5;
   // btc_delegators contains all the btc delegators with the associated finality provider.
   repeated BTCDelegator btc_delegators = 6;
   // all the events and its indexes.
   repeated EventIndex events = 7;
-  // vp_dst_cache is the table of all providers voting power with the total at one specific block.
-  // TODO: remove this after not storing in the keeper store it anymore.
-  repeated VotingPowerDistCacheBlkHeight vp_dst_cache = 8;
-}
-
-// VotingPowerFP contains the information about the voting power
-// of an finality provider in a specific block height.
-message VotingPowerFP {
-  // block_height is the height of the block the voting power was stored.
-  uint64 block_height = 1;
-  // fp_btc_pk the finality provider btc public key.
-  bytes fp_btc_pk = 2 [ (gogoproto.customtype) = "github.com/babylonlabs-io/babylon/types.BIP340PubKey" ];
-  // voting_power is the power of the finality provider at this specific block height.
-  uint64 voting_power = 3;
-}
-
-// VotingPowerDistCacheBlkHeight the total voting power of the finality providers at one specific block height
-message VotingPowerDistCacheBlkHeight {
-  // block_height is the height of the block the voting power distribution cached was stored.
-  uint64 block_height = 1;
-  // vp_distribution the finality providers distribution cache at that height.
-  VotingPowerDistCache vp_distribution = 2;
 }
 
 // BlockHeightBbnToBtc stores the btc <-> bbn block.
