+++ test/e2e/configurer/chain/queries_btcstaking.go
@@ -45,12 +45,12 @@ func (n *NodeConfig) QueryFinalityProviders() []*bstypes.FinalityProviderRespons
 	return resp.FinalityProviders
 }
 
-func (n *NodeConfig) QueryActiveFinalityProvidersAtHeight(height uint64) []*bstypes.ActiveFinalityProvidersAtHeightResponse {
-	path := fmt.Sprintf("/babylon/btcstaking/v1/finality_providers/%d", height)
+func (n *NodeConfig) QueryActiveFinalityProvidersAtHeight(height uint64) []*ftypes.ActiveFinalityProvidersAtHeightResponse {
+	path := fmt.Sprintf("/babylon/finality/v1/finality_providers/%d", height)
 	bz, err := n.QueryGRPCGateway(path, url.Values{})
 	require.NoError(n.t, err)
 
-	var resp bstypes.QueryActiveFinalityProvidersAtHeightResponse
+	var resp ftypes.QueryActiveFinalityProvidersAtHeightResponse
 	err = util.Cdc.UnmarshalJSON(bz, &resp)
 	require.NoError(n.t, err)
 
@@ -106,12 +106,12 @@ func (n *NodeConfig) QueryActiveDelegations() []*bstypes.BTCDelegationResponse {
 }
 
 func (n *NodeConfig) QueryActivatedHeight() (uint64, error) {
-	bz, err := n.QueryGRPCGateway("/babylon/btcstaking/v1/activated_height", url.Values{})
+	bz, err := n.QueryGRPCGateway("/babylon/finality/v1/activated_height", url.Values{})
 	if err != nil {
 		return 0, err
 	}
 
-	var resp bstypes.QueryActivatedHeightResponse
+	var resp ftypes.QueryActivatedHeightResponse
 	err = util.Cdc.UnmarshalJSON(bz, &resp)
 	if err != nil {
 		return 0, err
@@ -121,7 +121,6 @@ func (n *NodeConfig) QueryActivatedHeight() (uint64, error) {
 }
 
 // TODO: pagination support
-// TODO: remove public randomness storage?
 func (n *NodeConfig) QueryListPublicRandomness(fpBTCPK *bbn.BIP340PubKey) map[uint64]*bbn.SchnorrPubRand {
 	path := fmt.Sprintf("/babylon/finality/v1/finality_providers/%s/public_randomness_list", fpBTCPK.MarshalHex())
 	bz, err := n.QueryGRPCGateway(path, url.Values{})
