+++ testutil/datagen/incentive.go
@@ -7,7 +7,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	btcctypes "github.com/babylonlabs-io/babylon/x/btccheckpoint/types"
-	bstypes "github.com/babylonlabs-io/babylon/x/btcstaking/types"
+	ftypes "github.com/babylonlabs-io/babylon/x/finality/types"
 	itypes "github.com/babylonlabs-io/babylon/x/incentive/types"
 )
 
@@ -77,26 +77,26 @@ func GenRandomGauge(r *rand.Rand) *itypes.Gauge {
 	return itypes.NewGauge(coins...)
 }
 
-func GenRandomBTCDelDistInfo(r *rand.Rand) (*bstypes.BTCDelDistInfo, error) {
+func GenRandomBTCDelDistInfo(r *rand.Rand) (*ftypes.BTCDelDistInfo, error) {
 	btcPK, err := GenRandomBIP340PubKey(r)
 	if err != nil {
 		return nil, err
 	}
-	return &bstypes.BTCDelDistInfo{
+	return &ftypes.BTCDelDistInfo{
 		BtcPk:      btcPK,
 		StakerAddr: GenRandomAccount().Address,
 		TotalSat:   RandomInt(r, 1000) + 1,
 	}, nil
 }
 
-func GenRandomFinalityProviderDistInfo(r *rand.Rand) (*bstypes.FinalityProviderDistInfo, error) {
+func GenRandomFinalityProviderDistInfo(r *rand.Rand) (*ftypes.FinalityProviderDistInfo, error) {
 	// create finality provider with random commission
 	fp, err := GenRandomFinalityProvider(r)
 	if err != nil {
 		return nil, err
 	}
 	// create finality provider distribution info
-	fpDistInfo := bstypes.NewFinalityProviderDistInfo(fp)
+	fpDistInfo := ftypes.NewFinalityProviderDistInfo(fp)
 	// add a random number of BTC delegation distribution info
 	numBTCDels := RandomInt(r, 100) + 1
 	for i := uint64(0); i < numBTCDels; i++ {
@@ -111,8 +111,8 @@ func GenRandomFinalityProviderDistInfo(r *rand.Rand) (*bstypes.FinalityProviderD
 	return fpDistInfo, nil
 }
 
-func GenRandomVotingPowerDistCache(r *rand.Rand, maxFPs uint32) (*bstypes.VotingPowerDistCache, error) {
-	dc := bstypes.NewVotingPowerDistCache()
+func GenRandomVotingPowerDistCache(r *rand.Rand, maxFPs uint32) (*ftypes.VotingPowerDistCache, error) {
+	dc := ftypes.NewVotingPowerDistCache()
 	// a random number of finality providers
 	numFps := RandomInt(r, 10) + 1
 	for i := uint64(0); i < numFps; i++ {
