+++ x/btcstaking/keeper/grpc_query_test.go
@@ -26,31 +26,6 @@ import (
 
 var net = &chaincfg.SimNetParams
 
-func FuzzActivatedHeight(f *testing.F) {
-	datagen.AddRandomSeedsToFuzzer(f, 10)
-	f.Fuzz(func(t *testing.T, seed int64) {
-		r := rand.New(rand.NewSource(seed))
-
-		// Setup keeper and context
-		keeper, ctx := testkeeper.BTCStakingKeeper(t, nil, nil, nil)
-		ctx = sdk.UnwrapSDKContext(ctx)
-
-		// not activated yet
-		_, err := keeper.GetBTCStakingActivatedHeight(ctx)
-		require.Error(t, err)
-
-		randomActivatedHeight := datagen.RandomInt(r, 100) + 1
-		fp, err := datagen.GenRandomFinalityProvider(r)
-		require.NoError(t, err)
-		keeper.SetVotingPower(ctx, fp.BtcPk.MustMarshal(), randomActivatedHeight, uint64(10))
-
-		// now it's activated
-		resp, err := keeper.ActivatedHeight(ctx, &types.QueryActivatedHeightRequest{})
-		require.NoError(t, err)
-		require.Equal(t, randomActivatedHeight, resp.Height)
-	})
-}
-
 func FuzzFinalityProviders(f *testing.F) {
 	datagen.AddRandomSeedsToFuzzer(f, 10)
 	f.Fuzz(func(t *testing.T, seed int64) {
@@ -167,339 +142,6 @@ func FuzzFinalityProvider(f *testing.F) {
 	})
 }
 
-func FuzzPendingBTCDelegations(f *testing.F) {
-	datagen.AddRandomSeedsToFuzzer(f, 10)
-	f.Fuzz(func(t *testing.T, seed int64) {
-		r := rand.New(rand.NewSource(seed))
-		ctrl := gomock.NewController(t)
-		defer ctrl.Finish()
-
-		// Setup keeper and context
-		btclcKeeper := types.NewMockBTCLightClientKeeper(ctrl)
-		btccKeeper := types.NewMockBtcCheckpointKeeper(ctrl)
-		btccKeeper.EXPECT().GetParams(gomock.Any()).Return(btcctypes.DefaultParams()).AnyTimes()
-		keeper, ctx := testkeeper.BTCStakingKeeper(t, btclcKeeper, btccKeeper, nil)
-
-		// covenant and slashing addr
-		covenantSKs, covenantPKs, covenantQuorum := datagen.GenCovenantCommittee(r)
-		slashingAddress, err := datagen.GenRandomBTCAddress(r, net)
-		require.NoError(t, err)
-		slashingPkScript, err := txscript.PayToAddrScript(slashingAddress)
-		require.NoError(t, err)
-		slashingChangeLockTime := uint16(101)
-
-		// Generate a slashing rate in the range [0.1, 0.50] i.e., 10-50%.
-		// NOTE - if the rate is higher or lower, it may produce slashing or change outputs
-		// with value below the dust threshold, causing test failure.
-		// Our goal is not to test failure due to such extreme cases here;
-		// this is already covered in FuzzGeneratingValidStakingSlashingTx
-		slashingRate := sdkmath.LegacyNewDecWithPrec(int64(datagen.RandomInt(r, 41)+10), 2)
-
-		// Generate a random number of finality providers
-		numFps := datagen.RandomInt(r, 5) + 1
-		fps := []*types.FinalityProvider{}
-		for i := uint64(0); i < numFps; i++ {
-			fp, err := datagen.GenRandomFinalityProvider(r)
-			require.NoError(t, err)
-			AddFinalityProvider(t, ctx, *keeper, fp)
-			fps = append(fps, fp)
-		}
-
-		// Generate a random number of BTC delegations under each finality provider
-		startHeight := uint32(datagen.RandomInt(r, 100)) + 1
-		btclcKeeper.EXPECT().GetTipInfo(gomock.Any()).Return(&btclctypes.BTCHeaderInfo{Height: startHeight}).AnyTimes()
-
-		endHeight := uint32(datagen.RandomInt(r, 1000)) + startHeight + btcctypes.DefaultParams().CheckpointFinalizationTimeout + 1
-		stakingTime := endHeight - startHeight
-		numBTCDels := datagen.RandomInt(r, 10) + 1
-		pendingBtcDelsMap := make(map[string]*types.BTCDelegation)
-		for _, fp := range fps {
-			for j := uint64(0); j < numBTCDels; j++ {
-				delSK, _, err := datagen.GenRandomBTCKeyPair(r)
-				require.NoError(t, err)
-				// 0.5 chance that the delegation is created via pre-approval flow
-				if r.Intn(2) == 0 {
-					startHeight, endHeight = 0, 0
-				}
-				btcDel, err := datagen.GenRandomBTCDelegation(
-					r,
-					t,
-					net,
-					[]bbn.BIP340PubKey{*fp.BtcPk},
-					delSK,
-					covenantSKs,
-					covenantPKs,
-					covenantQuorum,
-					slashingPkScript,
-					stakingTime, startHeight, endHeight, 10000,
-					slashingRate,
-					slashingChangeLockTime,
-				)
-				require.NoError(t, err)
-				if datagen.RandomInt(r, 2) == 1 {
-					// remove covenant sig in random BTC delegations to make them inactive
-					btcDel.CovenantSigs = nil
-					pendingBtcDelsMap[btcDel.BtcPk.MarshalHex()] = btcDel
-				}
-				err = keeper.AddBTCDelegation(ctx, btcDel, btcDel.UnbondingTime-1)
-				require.NoError(t, err)
-
-				txHash := btcDel.MustGetStakingTxHash().String()
-				delView, err := keeper.BTCDelegation(ctx, &types.QueryBTCDelegationRequest{
-					StakingTxHashHex: txHash,
-				})
-				require.NoError(t, err)
-				require.NotNil(t, delView)
-			}
-		}
-
-		babylonHeight := datagen.RandomInt(r, 10) + 1
-		ctx = datagen.WithCtxHeight(ctx, babylonHeight)
-
-		// querying paginated BTC delegations and assert
-		// Generate a page request with a limit and a nil key
-		if len(pendingBtcDelsMap) == 0 {
-			return
-		}
-		limit := datagen.RandomInt(r, len(pendingBtcDelsMap)) + 1
-		pagination := constructRequestWithLimit(r, limit)
-		req := &types.QueryBTCDelegationsRequest{
-			Status:     types.BTCDelegationStatus_PENDING,
-			Pagination: pagination,
-		}
-		for i := uint64(0); i < numBTCDels; i += limit {
-			resp, err := keeper.BTCDelegations(ctx, req)
-			require.NoError(t, err)
-			require.NotNil(t, resp)
-			for _, btcDel := range resp.BtcDelegations {
-				_, ok := pendingBtcDelsMap[btcDel.BtcPk.MarshalHex()]
-				require.True(t, ok)
-				require.Equal(t, stakingTime, btcDel.StakingTime)
-			}
-			// Construct the next page request
-			pagination.Key = resp.Pagination.NextKey
-		}
-	})
-}
-
-func FuzzFinalityProviderPowerAtHeight(f *testing.F) {
-	datagen.AddRandomSeedsToFuzzer(f, 10)
-	f.Fuzz(func(t *testing.T, seed int64) {
-		r := rand.New(rand.NewSource(seed))
-
-		// Setup keeper and context
-		keeper, ctx := testkeeper.BTCStakingKeeper(t, nil, nil, nil)
-
-		// random finality provider
-		fp, err := datagen.GenRandomFinalityProvider(r)
-		require.NoError(t, err)
-		// add this finality provider
-		AddFinalityProvider(t, ctx, *keeper, fp)
-		// set random voting power at random height
-		randomHeight := datagen.RandomInt(r, 100) + 1
-		randomPower := datagen.RandomInt(r, 100) + 1
-		keeper.SetVotingPower(ctx, fp.BtcPk.MustMarshal(), randomHeight, randomPower)
-
-		// happy case
-		req1 := &types.QueryFinalityProviderPowerAtHeightRequest{
-			FpBtcPkHex: fp.BtcPk.MarshalHex(),
-			Height:     randomHeight,
-		}
-		resp, err := keeper.FinalityProviderPowerAtHeight(ctx, req1)
-		require.NoError(t, err)
-		require.Equal(t, randomPower, resp.VotingPower)
-
-		// case where the voting power store is not updated in
-		// the given height
-		requestHeight := randomHeight + datagen.RandomInt(r, 10) + 1
-		req2 := &types.QueryFinalityProviderPowerAtHeightRequest{
-			FpBtcPkHex: fp.BtcPk.MarshalHex(),
-			Height:     requestHeight,
-		}
-		_, err = keeper.FinalityProviderPowerAtHeight(ctx, req2)
-		require.ErrorIs(t, err, types.ErrVotingPowerTableNotUpdated)
-
-		// case where the given fp pk does not exist
-		randPk, err := datagen.GenRandomBIP340PubKey(r)
-		require.NoError(t, err)
-		req3 := &types.QueryFinalityProviderPowerAtHeightRequest{
-			FpBtcPkHex: randPk.MarshalHex(),
-			Height:     randomHeight,
-		}
-		_, err = keeper.FinalityProviderPowerAtHeight(ctx, req3)
-		require.ErrorIs(t, err, types.ErrFpNotFound)
-	})
-}
-
-func FuzzFinalityProviderCurrentVotingPower(f *testing.F) {
-	datagen.AddRandomSeedsToFuzzer(f, 10)
-	f.Fuzz(func(t *testing.T, seed int64) {
-		r := rand.New(rand.NewSource(seed))
-
-		// Setup keeper and context
-		keeper, ctx := testkeeper.BTCStakingKeeper(t, nil, nil, nil)
-
-		// random finality provider
-		fp, err := datagen.GenRandomFinalityProvider(r)
-		require.NoError(t, err)
-		// add this finality provider
-		AddFinalityProvider(t, ctx, *keeper, fp)
-		// set random voting power at random height
-		randomHeight := datagen.RandomInt(r, 100) + 1
-		ctx = datagen.WithCtxHeight(ctx, randomHeight)
-		randomPower := datagen.RandomInt(r, 100) + 1
-		keeper.SetVotingPower(ctx, fp.BtcPk.MustMarshal(), randomHeight, randomPower)
-
-		// assert voting power at current height
-		req := &types.QueryFinalityProviderCurrentPowerRequest{
-			FpBtcPkHex: fp.BtcPk.MarshalHex(),
-		}
-		resp, err := keeper.FinalityProviderCurrentPower(ctx, req)
-		require.NoError(t, err)
-		require.Equal(t, randomHeight, resp.Height)
-		require.Equal(t, randomPower, resp.VotingPower)
-
-		// if height increments but voting power hasn't recorded yet, then
-		// we need to return the height and voting power at the last height
-		ctx = datagen.WithCtxHeight(ctx, randomHeight+1)
-		resp, err = keeper.FinalityProviderCurrentPower(ctx, req)
-		require.NoError(t, err)
-		require.Equal(t, randomHeight, resp.Height)
-		require.Equal(t, randomPower, resp.VotingPower)
-
-		// test the case when the finality provider has 0 voting power
-		ctx = datagen.WithCtxHeight(ctx, randomHeight+2)
-		keeper.SetVotingPower(ctx, fp.BtcPk.MustMarshal(), randomHeight+2, 0)
-		resp, err = keeper.FinalityProviderCurrentPower(ctx, req)
-		require.NoError(t, err)
-		require.Equal(t, randomHeight+2, resp.Height)
-		require.Equal(t, uint64(0), resp.VotingPower)
-	})
-}
-
-func FuzzActiveFinalityProvidersAtHeight(f *testing.F) {
-	datagen.AddRandomSeedsToFuzzer(f, 10)
-	f.Fuzz(func(t *testing.T, seed int64) {
-		r := rand.New(rand.NewSource(seed))
-		ctrl := gomock.NewController(t)
-		defer ctrl.Finish()
-
-		// Setup keeper and context
-		btclcKeeper := types.NewMockBTCLightClientKeeper(ctrl)
-		btclcKeeper.EXPECT().GetTipInfo(gomock.Any()).Return(&btclctypes.BTCHeaderInfo{Height: 10}).AnyTimes()
-		btccKeeper := types.NewMockBtcCheckpointKeeper(ctrl)
-		btccKeeper.EXPECT().GetParams(gomock.Any()).Return(btcctypes.DefaultParams()).AnyTimes()
-		keeper, ctx := testkeeper.BTCStakingKeeper(t, btclcKeeper, btccKeeper, nil)
-
-		// covenant and slashing addr
-		covenantSKs, covenantPKs, covenantQuorum := datagen.GenCovenantCommittee(r)
-		slashingAddress, err := datagen.GenRandomBTCAddress(r, net)
-		require.NoError(t, err)
-		slashingPkScript, err := txscript.PayToAddrScript(slashingAddress)
-		require.NoError(t, err)
-		slashingChangeLockTime := uint16(101)
-
-		// Generate a slashing rate in the range [0.1, 0.50] i.e., 10-50%.
-		// NOTE - if the rate is higher or lower, it may produce slashing or change outputs
-		// with value below the dust threshold, causing test failure.
-		// Our goal is not to test failure due to such extreme cases here;
-		// this is already covered in FuzzGeneratingValidStakingSlashingTx
-		slashingRate := sdkmath.LegacyNewDecWithPrec(int64(datagen.RandomInt(r, 41)+10), 2)
-
-		// Generate a random batch of finality providers
-		var fps []*types.FinalityProvider
-		numFpsWithVotingPower := datagen.RandomInt(r, 10) + 1
-		numFps := numFpsWithVotingPower + datagen.RandomInt(r, 10)
-		for i := uint64(0); i < numFps; i++ {
-			fp, err := datagen.GenRandomFinalityProvider(r)
-			require.NoError(t, err)
-			AddFinalityProvider(t, ctx, *keeper, fp)
-			fps = append(fps, fp)
-		}
-
-		// For numFpsWithVotingPower finality providers, generate a random number of BTC delegations
-		numBTCDels := datagen.RandomInt(r, 10) + 1
-		babylonHeight := datagen.RandomInt(r, 10) + 1
-		fpsWithVotingPowerMap := make(map[string]*types.FinalityProvider)
-		for i := uint64(0); i < numFpsWithVotingPower; i++ {
-			fpBTCPK := fps[i].BtcPk
-			fpsWithVotingPowerMap[fpBTCPK.MarshalHex()] = fps[i]
-
-			var totalVotingPower uint64
-			for j := uint64(0); j < numBTCDels; j++ {
-				delSK, _, err := datagen.GenRandomBTCKeyPair(r)
-				require.NoError(t, err)
-				startHeight, endHeight := uint32(1), uint32(1000)
-				stakingTime := endHeight - startHeight
-				btcDel, err := datagen.GenRandomBTCDelegation(
-					r,
-					t,
-					net,
-					[]bbn.BIP340PubKey{*fpBTCPK},
-					delSK,
-					covenantSKs,
-					covenantPKs,
-					covenantQuorum,
-					slashingPkScript,
-					stakingTime, 1, 1000, 10000,
-					slashingRate,
-					slashingChangeLockTime,
-				)
-				require.NoError(t, err)
-				err = keeper.AddBTCDelegation(ctx, btcDel, btcDel.UnbondingTime-1)
-				require.NoError(t, err)
-				totalVotingPower += btcDel.TotalSat
-			}
-
-			keeper.SetVotingPower(ctx, fpBTCPK.MustMarshal(), babylonHeight, totalVotingPower)
-		}
-
-		// Test nil request
-		resp, err := keeper.ActiveFinalityProvidersAtHeight(ctx, nil)
-		if resp != nil {
-			t.Errorf("Nil input led to a non-nil response")
-		}
-		if err == nil {
-			t.Errorf("Nil input led to a nil error")
-		}
-
-		// Generate a page request with a limit and a nil key
-		limit := datagen.RandomInt(r, int(numFpsWithVotingPower)) + 1
-		pagination := constructRequestWithLimit(r, limit)
-		// Generate the initial query
-		req := types.QueryActiveFinalityProvidersAtHeightRequest{Height: babylonHeight, Pagination: pagination}
-		// Construct a mapping from the finality providers found to a boolean value
-		// Will be used later to evaluate whether all the finality providers were returned
-		fpsFound := make(map[string]bool, 0)
-
-		for i := uint64(0); i < numFpsWithVotingPower; i += limit {
-			resp, err = keeper.ActiveFinalityProvidersAtHeight(ctx, &req)
-			if err != nil {
-				t.Errorf("Valid request led to an error %s", err)
-			}
-			if resp == nil {
-				t.Fatalf("Valid request led to a nil response")
-			}
-
-			for _, fp := range resp.FinalityProviders {
-				// Check if the pk exists in the map
-				if _, ok := fpsWithVotingPowerMap[fp.BtcPkHex.MarshalHex()]; !ok {
-					t.Fatalf("rpc returned a finality provider that was not created")
-				}
-				fpsFound[fp.BtcPkHex.MarshalHex()] = true
-			}
-
-			// Construct the next page request
-			pagination = constructRequestWithKeyAndLimit(r, resp.Pagination.NextKey, limit)
-			req = types.QueryActiveFinalityProvidersAtHeightRequest{Height: babylonHeight, Pagination: pagination}
-		}
-
-		if len(fpsFound) != len(fpsWithVotingPowerMap) {
-			t.Errorf("Some finality providers were missed. Got %d while %d were expected", len(fpsFound), len(fpsWithVotingPowerMap))
-		}
-	})
-}
-
 func FuzzFinalityProviderDelegations(f *testing.F) {
 	datagen.AddRandomSeedsToFuzzer(f, 10)
 	f.Fuzz(func(t *testing.T, seed int64) {
@@ -613,6 +255,121 @@ func FuzzFinalityProviderDelegations(f *testing.F) {
 		}
 		require.Equal(t, len(btcDelsFound), len(expectedBtcDelsMap))
 
+	})
+}
+
+func FuzzPendingBTCDelegations(f *testing.F) {
+	datagen.AddRandomSeedsToFuzzer(f, 10)
+	f.Fuzz(func(t *testing.T, seed int64) {
+		r := rand.New(rand.NewSource(seed))
+		ctrl := gomock.NewController(t)
+		defer ctrl.Finish()
+
+		// Setup keeper and context
+		btclcKeeper := types.NewMockBTCLightClientKeeper(ctrl)
+		btccKeeper := types.NewMockBtcCheckpointKeeper(ctrl)
+		btccKeeper.EXPECT().GetParams(gomock.Any()).Return(btcctypes.DefaultParams()).AnyTimes()
+		keeper, ctx := testkeeper.BTCStakingKeeper(t, btclcKeeper, btccKeeper, nil)
+
+		// covenant and slashing addr
+		covenantSKs, covenantPKs, covenantQuorum := datagen.GenCovenantCommittee(r)
+		slashingAddress, err := datagen.GenRandomBTCAddress(r, net)
+		require.NoError(t, err)
+		slashingPkScript, err := txscript.PayToAddrScript(slashingAddress)
+		require.NoError(t, err)
+		slashingChangeLockTime := uint16(101)
+
+		// Generate a slashing rate in the range [0.1, 0.50] i.e., 10-50%.
+		// NOTE - if the rate is higher or lower, it may produce slashing or change outputs
+		// with value below the dust threshold, causing test failure.
+		// Our goal is not to test failure due to such extreme cases here;
+		// this is already covered in FuzzGeneratingValidStakingSlashingTx
+		slashingRate := sdkmath.LegacyNewDecWithPrec(int64(datagen.RandomInt(r, 41)+10), 2)
+
+		// Generate a random number of finality providers
+		numFps := datagen.RandomInt(r, 5) + 1
+		fps := []*types.FinalityProvider{}
+		for i := uint64(0); i < numFps; i++ {
+			fp, err := datagen.GenRandomFinalityProvider(r)
+			require.NoError(t, err)
+			AddFinalityProvider(t, ctx, *keeper, fp)
+			fps = append(fps, fp)
+		}
+
+		// Generate a random number of BTC delegations under each finality provider
+		startHeight := uint32(datagen.RandomInt(r, 100)) + 1
+		btclcKeeper.EXPECT().GetTipInfo(gomock.Any()).Return(&btclctypes.BTCHeaderInfo{Height: startHeight}).AnyTimes()
+
+		endHeight := uint32(datagen.RandomInt(r, 1000)) + startHeight + btcctypes.DefaultParams().CheckpointFinalizationTimeout + 1
+		stakingTime := endHeight - startHeight
+		numBTCDels := datagen.RandomInt(r, 10) + 1
+		pendingBtcDelsMap := make(map[string]*types.BTCDelegation)
+		for _, fp := range fps {
+			for j := uint64(0); j < numBTCDels; j++ {
+				delSK, _, err := datagen.GenRandomBTCKeyPair(r)
+				require.NoError(t, err)
+				// 0.5 chance that the delegation is created via pre-approval flow
+				if r.Intn(2) == 0 {
+					startHeight, endHeight = 0, 0
+				}
+				btcDel, err := datagen.GenRandomBTCDelegation(
+					r,
+					t,
+					net,
+					[]bbn.BIP340PubKey{*fp.BtcPk},
+					delSK,
+					covenantSKs,
+					covenantPKs,
+					covenantQuorum,
+					slashingPkScript,
+					stakingTime, startHeight, endHeight, 10000,
+					slashingRate,
+					slashingChangeLockTime,
+				)
+				require.NoError(t, err)
+				if datagen.RandomInt(r, 2) == 1 {
+					// remove covenant sig in random BTC delegations to make them inactive
+					btcDel.CovenantSigs = nil
+					pendingBtcDelsMap[btcDel.BtcPk.MarshalHex()] = btcDel
+				}
+				err = keeper.AddBTCDelegation(ctx, btcDel, btcDel.UnbondingTime-1)
+				require.NoError(t, err)
+
+				txHash := btcDel.MustGetStakingTxHash().String()
+				delView, err := keeper.BTCDelegation(ctx, &types.QueryBTCDelegationRequest{
+					StakingTxHashHex: txHash,
+				})
+				require.NoError(t, err)
+				require.NotNil(t, delView)
+			}
+		}
+
+		babylonHeight := datagen.RandomInt(r, 10) + 1
+		ctx = datagen.WithCtxHeight(ctx, babylonHeight)
+
+		// querying paginated BTC delegations and assert
+		// Generate a page request with a limit and a nil key
+		if len(pendingBtcDelsMap) == 0 {
+			return
+		}
+		limit := datagen.RandomInt(r, len(pendingBtcDelsMap)) + 1
+		pagination := constructRequestWithLimit(r, limit)
+		req := &types.QueryBTCDelegationsRequest{
+			Status:     types.BTCDelegationStatus_PENDING,
+			Pagination: pagination,
+		}
+		for i := uint64(0); i < numBTCDels; i += limit {
+			resp, err := keeper.BTCDelegations(ctx, req)
+			require.NoError(t, err)
+			require.NotNil(t, resp)
+			for _, btcDel := range resp.BtcDelegations {
+				_, ok := pendingBtcDelsMap[btcDel.BtcPk.MarshalHex()]
+				require.True(t, ok)
+				require.Equal(t, stakingTime, btcDel.StakingTime)
+			}
+			// Construct the next page request
+			pagination.Key = resp.Pagination.NextKey
+		}
 	})
 }
