+++ x/btcstaking/types/keys.go
@@ -15,12 +15,12 @@ const (
 )
 
 var (
-	ParamsKey               = []byte{0x01} // key prefix for the parameters
-	FinalityProviderKey     = []byte{0x02} // key prefix for the finality providers
-	BTCDelegatorKey         = []byte{0x03} // key prefix for the BTC delegators
-	BTCDelegationKey        = []byte{0x04} // key prefix for the BTC delegations
-	VotingPowerKey          = []byte{0x05} // key prefix for the voting power
-	BTCHeightKey            = []byte{0x06} // key prefix for the BTC heights
-	VotingPowerDistCacheKey = []byte{0x07} // key prefix for voting power distribution cache
-	PowerDistUpdateKey      = []byte{0x08} // key prefix for power distribution update events
+	ParamsKey           = []byte{0x01} // key prefix for the parameters
+	FinalityProviderKey = []byte{0x02} // key prefix for the finality providers
+	BTCDelegatorKey     = []byte{0x03} // key prefix for the BTC delegators
+	BTCDelegationKey    = []byte{0x04} // key prefix for the BTC delegations
+	// 0x05 was used for something else in the past
+	BTCHeightKey = []byte{0x06} // key prefix for the BTC heights
+	// 0x07 was used for something else in the past
+	PowerDistUpdateKey = []byte{0x08} // key prefix for power distribution update events
 )
