+++ x/finality/keeper/genesis.go
@@ -48,6 +48,14 @@ func (k Keeper) InitGenesis(ctx context.Context, gs types.GenesisState) error {
 		}
 	}
 
+	for _, fpVP := range gs.VotingPowers {
+		k.SetVotingPower(ctx, *fpVP.FpBtcPk, fpVP.BlockHeight, fpVP.VotingPower)
+	}
+
+	for _, vpCache := range gs.VpDstCache {
+		k.SetVotingPowerDistCache(ctx, vpCache.BlockHeight, vpCache.VpDistribution)
+	}
+
 	return k.SetParams(ctx, gs.Params)
 }
 
@@ -83,6 +91,16 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) 
 		return nil, err
 	}
 
+	vpFps, err := k.fpVotingPowers(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	vpDstCache, err := k.votingPowersDistCacheBlkHeight(ctx)
+	if err != nil {
+		return nil, err
+	}
+
 	return &types.GenesisState{
 		Params:           k.GetParams(ctx),
 		IndexedBlocks:    blocks,
@@ -92,6 +110,8 @@ func (k Keeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) 
 		PubRandCommit:    prCommit,
 		SigningInfos:     signingInfos,
 		MissedBlocks:     missedBlocks,
+		VotingPowers:     vpFps,
+		VpDstCache:       vpDstCache,
 	}, nil
 }
 
@@ -252,6 +272,49 @@ func (k Keeper) signingInfosAndMissedBlock(ctx context.Context) ([]types.Signing
 	}
 
 	return signingInfos, missedBlocks, nil
+}
+
+// fpVotingPowers gets the voting power of a given finality provider at a given Babylon height.
+func (k Keeper) fpVotingPowers(ctx context.Context) ([]*types.VotingPowerFP, error) {
+	iter := k.votingPowerStore(ctx).Iterator(nil, nil)
+	defer iter.Close()
+
+	vpFps := make([]*types.VotingPowerFP, 0)
+
+	for ; iter.Valid(); iter.Next() {
+		blkHeight, fpBTCPK, err := btcstk.ParseBlkHeightAndPubKeyFromStoreKey(iter.Key())
+		if err != nil {
+			return nil, err
+		}
+
+		vp := sdk.BigEndianToUint64(iter.Value())
+		vpFps = append(vpFps, &types.VotingPowerFP{
+			BlockHeight: blkHeight,
+			FpBtcPk:     fpBTCPK,
+			VotingPower: vp,
+		})
+	}
+
+	return vpFps, nil
+}
+
+func (k Keeper) votingPowersDistCacheBlkHeight(ctx context.Context) ([]*types.VotingPowerDistCacheBlkHeight, error) {
+	vps := make([]*types.VotingPowerDistCacheBlkHeight, 0)
+	iter := k.votingPowerDistCacheStore(ctx).Iterator(nil, nil)
+	defer iter.Close()
+
+	for ; iter.Valid(); iter.Next() {
+		var dc types.VotingPowerDistCache
+		if err := dc.Unmarshal(iter.Value()); err != nil {
+			return nil, err
+		}
+		vps = append(vps, &types.VotingPowerDistCacheBlkHeight{
+			BlockHeight:    sdk.BigEndianToUint64(iter.Key()),
+			VpDistribution: &dc,
+		})
+	}
+
+	return vps, nil
 }
 
 // parsePubKeyAndBlkHeightFromStoreKey expects to receive a key with
