+++ x/finality/keeper/power_table_test.go
@@ -12,8 +12,159 @@ import (
 	"github.com/babylonlabs-io/babylon/testutil/datagen"
 	btclctypes "github.com/babylonlabs-io/babylon/x/btclightclient/types"
 	"github.com/babylonlabs-io/babylon/x/btcstaking/types"
+	ftypes "github.com/babylonlabs-io/babylon/x/finality/types"
 )
 
+func FuzzVotingPowerTable(f *testing.F) {
+	datagen.AddRandomSeedsToFuzzer(f, 10)
+
+	f.Fuzz(func(t *testing.T, seed int64) {
+		r := rand.New(rand.NewSource(seed))
+		ctrl := gomock.NewController(t)
+		defer ctrl.Finish()
+
+		// mock BTC light client and BTC checkpoint modules
+		btclcKeeper := types.NewMockBTCLightClientKeeper(ctrl)
+		btccKeeper := types.NewMockBtcCheckpointKeeper(ctrl)
+		h := testutil.NewHelper(t, btclcKeeper, btccKeeper)
+
+		// set all parameters
+		covenantSKs, _ := h.GenAndApplyParams(r)
+		changeAddress, err := datagen.GenRandomBTCAddress(r, h.Net)
+		h.NoError(err)
+
+		// generate a random batch of finality providers, and commit pub rand list with timestamp
+		fps := []*types.FinalityProvider{}
+		numFpsWithVotingPower := datagen.RandomInt(r, 10) + 2
+		numFps := numFpsWithVotingPower + datagen.RandomInt(r, 10)
+		for i := uint64(0); i < numFps; i++ {
+			fpSK, _, fp := h.CreateFinalityProvider(r)
+			h.CommitPubRandList(r, fpSK, fp, 1, 100, true)
+			fps = append(fps, fp)
+		}
+
+		// for the first numFpsWithVotingPower finality providers, generate a random number of BTC delegations
+		numBTCDels := datagen.RandomInt(r, 10) + 1
+		stakingValue := datagen.RandomInt(r, 100000) + 100000
+		for i := uint64(0); i < numFpsWithVotingPower; i++ {
+			for j := uint64(0); j < numBTCDels; j++ {
+				delSK, _, err := datagen.GenRandomBTCKeyPair(r)
+				h.NoError(err)
+				stakingTxHash, delMsg, del, btcHeaderInfo, inclusionProof, _, err := h.CreateDelegation(
+					r,
+					delSK,
+					fps[i].BtcPk.MustToBTCPK(),
+					changeAddress.EncodeAddress(),
+					int64(stakingValue),
+					1000,
+					0,
+					0,
+					true,
+				)
+				h.NoError(err)
+				h.CreateCovenantSigs(r, covenantSKs, delMsg, del)
+				h.AddInclusionProof(stakingTxHash, btcHeaderInfo, inclusionProof)
+			}
+		}
+
+		/*
+			assert the first numFpsWithVotingPower finality providers have voting power
+		*/
+		babylonHeight := datagen.RandomInt(r, 10) + 1
+		h.SetCtxHeight(babylonHeight)
+		h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(&btclctypes.BTCHeaderInfo{Height: 30}).AnyTimes()
+		err = h.BTCStakingKeeper.BeginBlocker(h.Ctx)
+		require.NoError(t, err)
+		err = h.FinalityKeeper.BeginBlocker(h.Ctx)
+		require.NoError(t, err)
+
+		for i := uint64(0); i < numFpsWithVotingPower; i++ {
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
+			require.Equal(t, numBTCDels*stakingValue, power)
+		}
+		for i := numFpsWithVotingPower; i < numFps; i++ {
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
+			require.Zero(t, power)
+		}
+
+		// also, get voting power table and assert consistency
+		powerTable := h.FinalityKeeper.GetVotingPowerTable(h.Ctx, babylonHeight)
+		require.NotNil(t, powerTable)
+		for i := uint64(0); i < numFpsWithVotingPower; i++ {
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
+			require.Equal(t, powerTable[fps[i].BtcPk.MarshalHex()], power)
+		}
+		// the activation height should be the current Babylon height as well
+		activatedHeight, err := h.FinalityKeeper.GetBTCStakingActivatedHeight(h.Ctx)
+		require.NoError(t, err)
+		require.Equal(t, babylonHeight, activatedHeight)
+
+		/*
+			slash a random finality provider and move on
+			then assert the slashed finality provider does not have voting power
+		*/
+		// move to next Babylon height
+		h.BTCLightClientKeeper = btclcKeeper
+		babylonHeight += 1
+		h.SetCtxHeight(babylonHeight)
+		h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(&btclctypes.BTCHeaderInfo{Height: 30}).AnyTimes()
+		// slash a random finality provider
+		slashedIdx := datagen.RandomInt(r, int(numFpsWithVotingPower))
+		slashedFp := fps[slashedIdx]
+		err = h.BTCStakingKeeper.SlashFinalityProvider(h.Ctx, slashedFp.BtcPk.MustMarshal())
+		require.NoError(t, err)
+		// index height and record power table
+		err = h.BTCStakingKeeper.BeginBlocker(h.Ctx)
+		require.NoError(t, err)
+		err = h.FinalityKeeper.BeginBlocker(h.Ctx)
+		require.NoError(t, err)
+
+		// check if the slashed finality provider's voting power becomes zero
+		for i := uint64(0); i < numFpsWithVotingPower; i++ {
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
+			if i == slashedIdx {
+				require.Zero(t, power)
+			} else {
+				require.Equal(t, numBTCDels*stakingValue, power)
+			}
+		}
+		for i := numFpsWithVotingPower; i < numFps; i++ {
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
+			require.Zero(t, power)
+		}
+
+		// also, get voting power table and assert consistency
+		powerTable = h.FinalityKeeper.GetVotingPowerTable(h.Ctx, babylonHeight)
+		require.NotNil(t, powerTable)
+		for i := uint64(0); i < numFpsWithVotingPower; i++ {
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fps[i].BtcPk, babylonHeight)
+			if i == slashedIdx {
+				require.Zero(t, power)
+			}
+			require.Equal(t, powerTable[fps[i].BtcPk.MarshalHex()], power)
+		}
+
+		/*
+			move to 999th BTC block, then assert none of finality providers has voting power (since end height - w < BTC height)
+		*/
+		// replace the old mocked keeper
+		babylonHeight += 1
+		h.SetCtxHeight(babylonHeight)
+		h.BTCLightClientKeeper.EXPECT().GetTipInfo(gomock.Eq(h.Ctx)).Return(&btclctypes.BTCHeaderInfo{Height: 999}).AnyTimes()
+		err = h.BTCStakingKeeper.BeginBlocker(h.Ctx)
+		require.NoError(t, err)
+
+		for _, fp := range fps {
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, *fp.BtcPk, babylonHeight)
+			require.Zero(t, power)
+		}
+
+		// the activation height should be same as before
+		activatedHeight2, err := h.FinalityKeeper.GetBTCStakingActivatedHeight(h.Ctx)
+		require.NoError(t, err)
+		require.Equal(t, activatedHeight, activatedHeight2)
+	})
+}
 func FuzzRecordVotingPowerDistCache(f *testing.F) {
 	datagen.AddRandomSeedsToFuzzer(f, 10)
 
@@ -78,7 +229,7 @@ func FuzzRecordVotingPowerDistCache(f *testing.F) {
 		h.BeginBlocker()
 
 		// assert voting power distribution cache is correct
-		dc := h.BTCStakingKeeper.GetVotingPowerDistCache(h.Ctx, babylonHeight)
+		dc := h.FinalityKeeper.GetVotingPowerDistCache(h.Ctx, babylonHeight)
 		require.NotNil(t, dc)
 		require.Equal(t, dc.TotalBondedSat, numFpsWithVotingPower*numBTCDels*stakingValue, dc.String())
 		activeFPs := dc.GetActiveFinalityProviderSet()
@@ -114,7 +265,7 @@ func FuzzVotingPowerTable_ActiveFinalityProviders(f *testing.F) {
 		h.NoError(err)
 
 		// generate a random batch of finality providers, each with a BTC delegation with random power
-		fpsWithMeta := []*types.FinalityProviderDistInfo{}
+		fpsWithMeta := []*ftypes.FinalityProviderDistInfo{}
 		numFps := datagen.RandomInt(r, 300) + 1
 		noTimestampedFps := map[string]bool{}
 		for i := uint64(0); i < numFps; i++ {
@@ -142,7 +293,7 @@ func FuzzVotingPowerTable_ActiveFinalityProviders(f *testing.F) {
 
 			// 30 percent not have timestamped randomness, which causes
 			// zero voting power in the table
-			fpDistInfo := &types.FinalityProviderDistInfo{BtcPk: fp.BtcPk, TotalBondedSat: stakingValue}
+			fpDistInfo := &ftypes.FinalityProviderDistInfo{BtcPk: fp.BtcPk, TotalBondedSat: stakingValue}
 			if r.Intn(10) <= 2 {
 				h.CommitPubRandList(r, fpSK, fp, 1, 100, false)
 				noTimestampedFps[fp.BtcPk.MarshalHex()] = true
@@ -158,7 +309,7 @@ func FuzzVotingPowerTable_ActiveFinalityProviders(f *testing.F) {
 
 		maxActiveFpsParam := h.FinalityKeeper.GetParams(h.Ctx).MaxActiveFinalityProviders
 		// get a map of expected active finality providers
-		types.SortFinalityProvidersWithZeroedVotingPower(fpsWithMeta)
+		ftypes.SortFinalityProvidersWithZeroedVotingPower(fpsWithMeta)
 		expectedActiveFps := fpsWithMeta[:min(uint32(len(fpsWithMeta)-len(noTimestampedFps)), maxActiveFpsParam)]
 		expectedActiveFpsMap := map[string]uint64{}
 		for _, fp := range expectedActiveFps {
@@ -173,7 +324,7 @@ func FuzzVotingPowerTable_ActiveFinalityProviders(f *testing.F) {
 
 		// only finality providers in expectedActiveFpsMap have voting power
 		for _, fp := range fpsWithMeta {
-			power := h.BTCStakingKeeper.GetVotingPower(h.Ctx, fp.BtcPk.MustMarshal(), babylonHeight)
+			power := h.FinalityKeeper.GetVotingPower(h.Ctx, fp.BtcPk.MustMarshal(), babylonHeight)
 			if expectedPower, ok := expectedActiveFpsMap[fp.BtcPk.MarshalHex()]; ok {
 				require.Equal(t, expectedPower, power)
 			} else {
@@ -183,7 +334,7 @@ func FuzzVotingPowerTable_ActiveFinalityProviders(f *testing.F) {
 
 		// also, get voting power table and assert there is
 		// min(len(expectedActiveFps), MaxActiveFinalityProviders) active finality providers
-		powerTable := h.BTCStakingKeeper.GetVotingPowerTable(h.Ctx, babylonHeight)
+		powerTable := h.FinalityKeeper.GetVotingPowerTable(h.Ctx, babylonHeight)
 		expectedNumActiveFps := len(expectedActiveFpsMap)
 		if expectedNumActiveFps > int(maxActiveFpsParam) {
 			expectedNumActiveFps = int(maxActiveFpsParam)
@@ -274,11 +425,11 @@ func FuzzVotingPowerTable_ActiveFinalityProviderRotation(f *testing.F) {
 			return fpsWithMeta[i].VotingPower > fpsWithMeta[j].VotingPower
 		})
 		for i := 0; i < numActiveFPs; i++ {
-			votingPower := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
+			votingPower := h.FinalityKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
 			require.Equal(t, fpsWithMeta[i].VotingPower, votingPower)
 		}
 		for i := numActiveFPs; i < int(numFps); i++ {
-			votingPower := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
+			votingPower := h.FinalityKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
 			require.Zero(t, votingPower)
 		}
 
@@ -364,11 +515,11 @@ func FuzzVotingPowerTable_ActiveFinalityProviderRotation(f *testing.F) {
 			return fpsWithMeta[i].VotingPower > fpsWithMeta[j].VotingPower
 		})
 		for i := 0; i < numActiveFPs; i++ {
-			votingPower := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
+			votingPower := h.FinalityKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
 			require.Equal(t, fpsWithMeta[i].VotingPower, votingPower)
 		}
 		for i := numActiveFPs; i < int(numFps); i++ {
-			votingPower := h.BTCStakingKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
+			votingPower := h.FinalityKeeper.GetVotingPower(h.Ctx, *fpsWithMeta[i].BtcPk, babylonHeight)
 			require.Zero(t, votingPower)
 		}
 	})
