+++ x/finality/keeper/tallying.go
@@ -19,7 +19,7 @@ import (
 // but without block that has finality providers set AND does not receive QC
 func (k Keeper) TallyBlocks(ctx context.Context) {
 	sdkCtx := sdk.UnwrapSDKContext(ctx)
-	activatedHeight, err := k.BTCStakingKeeper.GetBTCStakingActivatedHeight(ctx)
+	activatedHeight, err := k.GetBTCStakingActivatedHeight(ctx)
 	if err != nil {
 		// invoking TallyBlocks when BTC staking protocol is not activated is a programming error
 		panic(fmt.Errorf("cannot tally a block when the BTC staking protocol hasn't been activated yet, current height: %v, activated height: %v",
@@ -46,7 +46,7 @@ func (k Keeper) TallyBlocks(ctx context.Context) {
 		}
 
 		// get the finality provider set of this block
-		fpSet := k.BTCStakingKeeper.GetVotingPowerTable(ctx, ib.Height)
+		fpSet := k.GetVotingPowerTable(ctx, ib.Height)
 
 		if fpSet != nil && !ib.Finalized {
 			// has finality providers, non-finalised: tally and try to finalise the block
@@ -84,7 +84,7 @@ func (k Keeper) finalizeBlock(ctx context.Context, block *types.IndexedBlock, vo
 	// set next height to finalise as height+1
 	k.setNextHeightToFinalize(ctx, block.Height+1)
 	// distribute rewards to BTC staking stakeholders w.r.t. the voting power distribution cache
-	dc := k.BTCStakingKeeper.GetVotingPowerDistCache(ctx, block.Height)
+	dc := k.GetVotingPowerDistCache(ctx, block.Height)
 	if dc == nil {
 		// failing to get a voting power distribution cache before distributing reward is a programming error
 		panic(fmt.Errorf("voting power distribution cache not found at height %d", block.Height))
@@ -94,7 +94,7 @@ func (k Keeper) finalizeBlock(ctx context.Context, block *types.IndexedBlock, vo
 	// reward voted finality providers
 	k.IncentiveKeeper.RewardBTCStaking(ctx, block.Height, filteredDc)
 	// remove reward distribution cache afterwards
-	k.BTCStakingKeeper.RemoveVotingPowerDistCache(ctx, block.Height)
+	k.RemoveVotingPowerDistCache(ctx, block.Height)
 	// record the last finalized height metric
 	types.RecordLastFinalizedHeight(block.Height)
 }
