+++ x/finality/types/errors.go
@@ -19,5 +19,7 @@ var (
 	ErrNoSlashableEvidence            = errorsmod.Register(ModuleName, 1110, "there is no slashable evidence")
 	ErrPubRandCommitNotBTCTimestamped = errorsmod.Register(ModuleName, 1111, "the public randomness commit is not BTC timestamped yet")
 	ErrJailingPeriodNotPassed         = errorsmod.Register(ModuleName, 1112, "the jailing period is not passed")
-	ErrFinalityNotActivated           = errorsmod.Register(ModuleName, 1113, "finality is not active yet")
+	ErrVotingPowerTableNotUpdated     = errorsmod.Register(ModuleName, 1113, "voting power table has not been updated")
+	ErrBTCStakingNotActivated         = errorsmod.Register(ModuleName, 1114, "the BTC staking protocol is not activated yet")
+	ErrFinalityNotActivated           = errorsmod.Register(ModuleName, 1115, "finality is not active yet")
 )
